Massive changes, but videos can now be linked into the gallery!
Modified the gallery-building logic to accept an arbritrary list instead of automatically building one from the directory.
This is primarily used to allow external videos (youtube only for now) to be linked, using the video's default thumbnail as the image, and upon click, opening a new window to the video.
( Embedding an iframe doesn't work with PhotoSwipe )

To use the override, the following format is supported:
---
params:
  gallery:
    - localImg1.jpg
    - localImg2.jpg
    - youtube <id>
    - yt <id>
---

Still todo:
- Add custom CSS/SVG as an overlay to the thumbnail to indicate that it's a video
- Should probably add edge cases to the externally linking JS